move types to .ts
