+++ src/main.ts
@@ -5,6 +5,9 @@
  * @module Main Functions
  */
 
+import type * as qasm2 from "./qasm2/ast"
+import type * as qasm3 from "./qasm3/ast"
+
 import * as fs from "fs";
 import { OpenQASMVersion, OpenQASMMajorVersion } from "./version";
 import { lex } from "./lexer";
@@ -61,6 +64,41 @@ import { parse } from "./parser";
  * // Output will include __className__ properties for each node
  * ```
  */
+
+export function parseString(
+  qasm: string,
+  version: number | OpenQASMVersion | OpenQASMMajorVersion,
+  verbose: boolean,
+  stringify: true
+): string;
+
+export function parseString(
+  qasm: string,
+  version:
+    2 |
+    (OpenQASMVersion & { major: OpenQASMMajorVersion.Version2 }) |
+    OpenQASMMajorVersion.Version2,
+  verbose?: boolean,
+  stringify?: boolean
+): qasm2.AstNode[];
+
+export function parseString(
+  qasm: string,
+  version:
+    3 |
+    (OpenQASMVersion & { major: OpenQASMMajorVersion.Version3 }) |
+    OpenQASMMajorVersion.Version3,
+  verbose?: boolean,
+  stringify?: boolean
+): qasm3.AstNode[];
+
+export function parseString(
+  qasm: string,
+  version?: number | OpenQASMVersion | OpenQASMMajorVersion,
+  verbose?: boolean,
+  stringify?: boolean
+): string | qasm2.AstNode[] | qasm3.AstNode[];
+
 export function parseString(
   qasm: string,
   version?: number | OpenQASMVersion | OpenQASMMajorVersion,
@@ -118,7 +156,7 @@ export function parseFile(
 ) {
   return parseString(
     fs.readFileSync(file, "utf8"),
-    version,
+    version as any,
     verbose,
     stringify,
   );
